watchdog/ml/requirements.txt
torch
transformers
Pillow
scikit-learn
pdf2image
numpy
joblib
